+++ activestorage/app/models/active_storage/variant_with_record.rb
@@ -3,7 +3,8 @@
 # = Active Storage \Variant With Record
 #
 # Like an ActiveStorage::Variant, but keeps detail about the variant in the database as an
-# ActiveStorage::VariantRecord. This is only used if +ActiveStorage.track_variants+ is enabled.
+# ActiveStorage::VariantRecord. This is only used if +ActiveStorage.track_variants+ is enabled
+# or the variant is generated with the option of +track: true+.
 class ActiveStorage::VariantWithRecord
   include ActiveStorage::Blob::Servable